requirements.txt
Flask==3.1.0
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
google-cloud-language==2.16.0
gunicorn==22.0.0

